added a new class InputControl
