+++ game/manager/Player.java
@@ -27,12 +27,6 @@ public class Player {
         this.lifePoints = lifePoints;
     }
 
-    //gets user input
-    public static String initializePlayer(){
-        Scanner scr=new Scanner(System.in);
-        String player= scr.nextLine();
-        System.out.println("Hi player! welcome to the game, what is your name? " + player );
-        return player;
-    }
+
 
 }
